refactor: Q클래스 import 경로 변경
